coherence-sre/sprint2/requirements_sprint2.txt
# Sprint 2: Correlation Engine Dependencies
# Python 3.9+ required

# Supabase client for PostgreSQL database operations
supabase==2.9.0

# Python-dotenv for loading environment variables from .env file
python-dotenv==1.0.1

# Pandas for time-series analysis and data manipulation
pandas==2.2.0

# NumPy for numerical operations (pandas dependency)
numpy==1.26.3
